Restructured testing directory
