+++ tests/models/test_model_options.py
